begin sculpting fish
implemented more transformation tools, including a fixed fill_ring_pole, and a tool that creates a spline out of extruded rings, with custom pos and scale for each loop. Also added a subdivided box function, which will be the basis for fish fins. Test geometry for the fish body currently appears in the scene alongside plants